Added Flag for loop support with custom start time
The flag is used to determine if loop video is set to true,
once the video starts, options_['loop'] will always be set to false.
This allows the onEnded event to always be fired and allows us to set
custom start-time for when the video loops.

(When loop is normally set the onEnd end onPlay events are never called
their for this flag is necessary)