Fix Player.buffered() to more correct behavior
The problem is that Player.buffered() always return the first time range. This means that while do seeking threough the timeline, progress loader is not updated as far as browser send new buffering info at the tail of buffered time-ranges, so, looks like it would be more correctly to grab this info from the last time-range object rather than from the first one.