Fixed vertical slider check when obfuscated
In #477, support for vertical slider computation was added.
The config check breaks when obfuscated, it ends up being something like `if(a.j.Wd)`

I've just started to familiarize myself with the code-base, but I believe `this.options().vertical` is the correct way to check an option. This should solve the obfuscation issue.