Add WOFF2 font
WOFF2 is a new font format optimized for size, currently supported in Blink stable:

http://caniuse.com/#feat=woff2

and FF nightlies:

https://bugzilla.mozilla.org/show_bug.cgi?id=1064737

File size savings of 30-50%. This is the future.

Decompression is expensive, but since the bottleneck is typically on the wire, this is still probably a good idea, since font loading block the UI. On this file, the WOFF vs WOFF2 file difference is 1.2K, tiny but still probably worth doing, as font loading blocks the UI render, and every byte counts on mobile.