Add Chinese Simplified localization file. Found one new string that may needs localization ("The video is encrypted and we do not have the keys to decrypt it.").
Was not able to run test on my Windows 8.1 computer.
When run grunt, it complains:
Warning: Command failed: 'java' is not recognized as an internal or external command.
When run grunt test, it complains:
PhantomJS 1.9.7 (Windows 8) Player Minified should be able to access expected player API methods FAILED.
Although the tests are failed, I don't see the localization has anything to do with the error.