Listen for posterchange on tech
When someone is using a YouTube video and doesn't have a poster, we need to put the YouTube default poster. However, this is an async process. We need to check if the highres version exists. If it doesn't we use the low res.

To make that possible, the player has to listen for posterchange on the tech and trigger it again on himself so that the poster can set himself correctly.