Roll back changes and remove breaking code
After some pretty rigorous testing today and discussion with @dmlap, it
seems the best course of action here is to let Safari on iOS keep its
default behavior because it appears that it caches caption language
preferences in the player application rather than the browser.

Safari's behavior appears to be:

- If no caption language was previously selected (e.g. "Off") then no
  captions are downloaded or displayed.
- When the user chooses a language, the .vtt file is downloaded and
  captions for that language begin.
- When the user leaves and returns to the video in question (even with
  an explicit cache clear) the previously selected .vtt file is used.

Removing this code lets all browsers/user agents do their default
behavior, which seems like the best course of action.