Consistent width for fullscreen button and time-control
This changes the width of the fullscreen button to 4em. For some
reason it was 3.8, but that doesn't make much sense and is not in line
with the other controls.

It also sets the textual time-control's width to auto with a 1em
padding to the left and right. Again this is more in line with the
behavior of the other controls.