Correct adaptive layout selectors to match their intent
There were a few problems
1: The fullscreen button was hidden in vjs-layout-x-small, unlike its
comments claimed and not matching with tiny, where it is visible.
2: The vjs-captions-control was hidden twice for tiny and x-small
3: The vjs-captions-control was hidden in small, even though the
intent was (per comments) for it to be visible
4: The vjs-volume-menu-button was present on x-small, because it was
incorrectly specified as .vjs-volume-button

All of this was partly because the lines for each layout were
different and thus 'hard' to compare between the layouts. I have now
split them in 4 groups/lines. Time controls, rate/progress, volume,
metadata. This will hopefully prevent similar problems in the future.