Add beforesourcechange and sourcechange events
These are non-standard events, but I'd like to open discussion on adding
them. They have a lot of practical application as we've seen at
Brightcove and other companies where core contributors work. This may be
ultimately more appropriate in a plugin, but modifying the `Player`
prototype seems out of the scope of plugins to me.