Add support for withCredentials on TextTracks
This change adds support for fetching text tracks that require
security cookies. The code checks for a crossorigin="use-credentials"
attribute on a given text track and, if defined, will set the
withCredentials option on the XHR fetch. We discovered this problem
when hosting closed caption data on AWS CloudFront which uses signed
cookies for access controls.