fix: prevent dupe events on enabled ClickableComponents (#4357)
Like #4316 for v6, this prevents multiple click handlers being added if an already enabled component is enabled. In the PR for six, ClickableComponent was tracking this.enabled_ but this is incompatible with v5 on a MenuButton which also changes the state of this.enabled_.
Removes event handlers in enable() before adding them.