fix: Safari picture-in-picture triggers fullscreenchange (#4442)
When picture-in-picture mode is entered on Safari, `webkitbeginfullscreen` is triggered which results in a proxied `fullscreenchange` event and adding the fullscreen class to the player. That causes the tech element to collapse to zero height so that the "this video is playing in picture in picture" placeholder is not visible.

As with #4437 for v6, on `webkitbeginfullscreen` check whether the presentation mode is picture-in-picture' before proxying a `fullscreenchange` event.