refactor: Internal "evented" mixin can better handle changing `el_`s.
In those rare cases where someone manually overwrites `el_` on a
component, current 6.x versions of Video.js will stop emitting events
properly for that component because it caches a reference to the `el_`
at init time. This refactor causes the evented mixin to cache the
_property key_ of the element instead of the element itself and finds
that element each time it's needed.