fix: Duration change fixed when duration is NaN
When duration is NaN, the update of the duration was being ignored.
I updated the duration(seconds) in player.js function to accept NaN values since the formatting of the time displayed already has exceptions for when the time passed is NaN. When this happens, the display of the duration is '-:-'.
I also added some lines to the documentation on the code that explains what I just described in this message.