fix: Changed the handling of the duration display update when the source changes in case preload='none'
When preload isn't available, you don't have access to the video data like the duration, so the duration changes to NaN by default. The duration change event is not triggered and the duration display doesn't change. If you set the duration to the new value (NaN) with the duration(seconds) function, the event is triggered.