Fix polymer issue by marking click event with stopPropagation
The play button stops working when recent versions of Google's Polymer
is in use on the page because of the way Polymer synthesizes 'tap'
events on non-touch devices. The Polymer tap code thinks the click
event was ignored unless the DOM event's stopPropagation method is
called.

Fixes: #5624 In chrome 70 the small play/pause control doesn't work with Polymer 1.x Gestures tap is used on document

Demo of issue: https://codepen.io/mscalora/pen/mQzQmp