Introduce write method to handle value output
I found myself needing to access the value that was being written to the time tooltip and the only way to do this was to overwrite the entire `update` method just to set the `time` content to the TimeTooltip state. This method would allow for overriding a single line of code if the DOM mutations aren't preferred.