fix: add use of the safe version of computedStyle
Add use of the safe version of computedStyle in the currentDimension method of Component's class.
This will prevent a null exception when a video.js is implemented in a 'display:none' iframe on Firefox (<62).

This is a fix in continuation of the MR#3664 regarding a bug in Firefox https://bugzilla.mozilla.org/show_bug.cgi?id=548397