lang: Clone zh-CN to zh-Hans and zh-TW to zh-Hant (#6098)
Currently, zh-CN and zh-TW are used as shorthands for Simplified and Traditional Chinese character sets, respectively. However, they carry with them a built-in assumption about the geographic locale which may not be appropriate in all cases.

The codes zh-Hans and zh-Hant are more correct in that they refer only to the character set and not the geographic locale. Further specificity is available via codes like zh-Hans-CN.