[#6436] - Create a helper to get focusable elements and trap focus in fullscreen mode to fix an a11y bug
As described in issue #6436, tabbing through the player's controlBar is not properly trapped when in fullscreen mode. This leads to a poor experience for those relying on screen-reader technology because their focus ends up being placed on an element that is outside of the player. This change ensures that tabbing on the last element in the controlBar is properly trapped such that the focus is placed on the first focusable element within the playerr